Added GitHub link to bottom.
Also added Social Media spot for more images. Format should remain the same for all of them, so no need to fidle with the css. 

Also, I did add one GitHub logo image from Wikimedia commons for the link. It's transparent. If you want to see the source you can do so in the last commit I made in this fork.

Awesome Sauce.